Обновил фильтрацию и дао, привел весь текст к одному виду

+++ src/main/java/hookah_sql/controller/TabaccoListController.java
@@ -6,16 +6,19 @@ import hookah_sql.dao.TabaccoDAO;
 import hookah_sql.tabacco.Tabacco;
 import hookah_sql.tabacco.TabaccoEnum;
 import hookah_sql.utils.Utils;
+import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.*;
 import org.springframework.stereotype.Component;
 
+import java.util.*;
+
 @Component("tabaccoListController")
 public class TabaccoListController implements Controller {
 
     @FXML
-    private TableView<Tabacco> tabaccoListTable;
+    protected TableView<Tabacco> tabaccoListTable;
 
     @FXML
     private TableColumn<Tabacco, String> nameColumn;
@@ -48,28 +51,22 @@ public class TabaccoListController implements Controller {
     private Label descriptionValue;
 
     @FXML
-    private Button allButton;
+    private Menu tabaccoFilterMenu;
 
     @FXML
-    private Button mustHaveButton;
+    private Menu hardnessFilterMenu;
 
     @FXML
-    private Button blackBurnButton;
+    private Menu tasteFilterMenu;
 
     @FXML
-    private Button satyrButton;
+    private Label tabaccoCount;
 
     @FXML
-    private Button darksideButton;
+    private Button resetButton;
 
     @FXML
-    private Button tangiersButton;
-
-    @FXML
-    private Button duftButton;
-
-    @FXML
-    private Button elementButton;
+    private Button clearButton;
 
     @FXML
     private TextField tabaccoAddNameField;
@@ -78,9 +75,9 @@ public class TabaccoListController implements Controller {
 
     @FXML
     private void initialize() {
-        refresh(TabaccoDAO.getTabaccoList(TabaccoEnum.ALL));
+        refresh();
         prepareTable();
-        prepareFilterButtons();
+        prepareFilterMenu();
     }
 
     protected void prepareTable() {
@@ -97,15 +94,18 @@ public class TabaccoListController implements Controller {
         );
     }
 
-    private void prepareFilterButtons() {
-        allButton.setOnAction(event -> refresh(TabaccoDAO.getTabaccoList(TabaccoEnum.ALL)));
-        blackBurnButton.setOnAction(event -> refresh(TabaccoDAO.getTabaccoList(TabaccoEnum.BLACKBURN)));
-        darksideButton.setOnAction(event -> refresh(TabaccoDAO.getTabaccoList(TabaccoEnum.DARKSIDE)));
-        duftButton.setOnAction(event -> refresh(TabaccoDAO.getTabaccoList(TabaccoEnum.DUFT)));
-        elementButton.setOnAction(event -> refresh(TabaccoDAO.getTabaccoList(TabaccoEnum.ELEMENT)));
-        mustHaveButton.setOnAction(event -> refresh(TabaccoDAO.getTabaccoList(TabaccoEnum.MUSTHAVE)));
-        satyrButton.setOnAction(event -> refresh(TabaccoDAO.getTabaccoList(TabaccoEnum.SATYR)));
-        tangiersButton.setOnAction(event -> refresh(TabaccoDAO.getTabaccoList(TabaccoEnum.TANGIERS)));
+    private void prepareFilterMenu() {
+        for (MenuItem menuItem : tabaccoFilterMenu.getItems()) {
+            menuItem.setOnAction(event -> refreshFromFilter((CheckMenuItem) menuItem));
+        }
+
+        for (MenuItem menuItem : tasteFilterMenu.getItems()) {
+            menuItem.setOnAction(event -> refreshFromFilter((CheckMenuItem) menuItem));
+        }
+
+        for (MenuItem menuItem : hardnessFilterMenu.getItems()) {
+            menuItem.setOnAction(event -> refreshFromFilter((CheckMenuItem) menuItem));
+        }
     }
 
     protected void showTabaccoDetails(Tabacco tabacco) {
@@ -126,9 +126,119 @@ public class TabaccoListController implements Controller {
         }
     }
 
-    protected void refresh(ObservableList<Tabacco> tabaccoObservableList) {
+    @FXML
+    private void refresh() {
         tabaccoListTable.refresh();
-        tabaccoListTable.setItems(tabaccoObservableList);
+
+        ObservableList<Tabacco> listToAdd = TabaccoDAO.getTabaccoList(TabaccoEnum.ALL);
+        tabaccoListTable.setItems(listToAdd);
+
+        cleanUpCheckBoxes(listToAdd);
+
+        tabaccoCount.setText(String.valueOf(listToAdd.size()));
+    }
+
+    private void refreshFromButtons() {
+        tabaccoListTable.refresh();
+
+        ObservableList<Tabacco> listToAdd = TabaccoDAO.getTabaccoList(
+                getAllChosenTabaccos(), getAllNotChosenTastes(), getAllChosenHardnesses()
+        );
+        tabaccoListTable.setItems(listToAdd);
+
+        cleanUpCheckBoxes(listToAdd);
+
+        tabaccoCount.setText(String.valueOf(listToAdd.size()));
+    }
+
+    private void refreshFromFilter(CheckMenuItem menuItem) {
+        tabaccoListTable.refresh();
+
+        ObservableList<Tabacco> listToAdd = TabaccoDAO.getTabaccoList(menuItem, tabaccoListTable.getItems()
+        );
+
+        tabaccoListTable.setItems(listToAdd);
+
+        cleanUpCheckBoxes(listToAdd);
+
+        tabaccoCount.setText(String.valueOf(listToAdd.size()));
+    }
+
+    private void cleanUpCheckBoxes(ObservableList<Tabacco> list) {
+        Set<String> allTabaccosInTable = new HashSet<>();
+        Set<String> allTastesInTable = new HashSet<>();
+        Set<Integer> allHardnessesInTable = new HashSet<>();
+
+        for (Tabacco tabacco : list) {
+            allTabaccosInTable.add(tabacco.getTabaccoName().toLowerCase());
+            allTastesInTable.addAll(Arrays.asList(tabacco.getTaste().split(", ")));
+            allHardnessesInTable.add(tabacco.getHardness());
+        }
+
+        for (MenuItem menuItem : tabaccoFilterMenu.getItems()) {
+            CheckMenuItem checkMenuItem = (CheckMenuItem) menuItem;
+            checkMenuItem.setSelected(allTabaccosInTable.contains(checkMenuItem.getText().toLowerCase()));
+        }
+
+        for (MenuItem menuItem : tasteFilterMenu.getItems()) {
+            CheckMenuItem checkMenuItem = (CheckMenuItem) menuItem;
+            checkMenuItem.setSelected(allTastesInTable.contains(checkMenuItem.getText()));
+        }
+
+        for (MenuItem menuItem : hardnessFilterMenu.getItems()) {
+            CheckMenuItem checkMenuItem = (CheckMenuItem) menuItem;
+            checkMenuItem.setSelected(allHardnessesInTable.contains(Integer.valueOf(checkMenuItem.getText())));
+        }
+    }
+
+    @FXML
+    private void clear() {
+        ObservableList<Tabacco> listToAdd = FXCollections.observableArrayList(new ArrayList<>());
+
+        tabaccoListTable.setItems(listToAdd);
+
+        cleanUpCheckBoxes(listToAdd);
+
+        tabaccoCount.setText(String.valueOf(listToAdd.size()));
+    }
+
+    private List<String> getAllChosenTabaccos() {
+        List<String> tabaccoNames = new ArrayList<>();
+
+        for (MenuItem menuItem : tabaccoFilterMenu.getItems()) {
+            CheckMenuItem checkMenuItem = (CheckMenuItem) menuItem;
+            if (checkMenuItem.isSelected()) {
+                tabaccoNames.add(checkMenuItem.getText());
+            }
+        }
+
+        return tabaccoNames;
+    }
+
+    private List<String> getAllNotChosenTastes() {
+        List<String> tastes = new ArrayList<>();
+
+        for (MenuItem menuItem : tasteFilterMenu.getItems()) {
+            CheckMenuItem checkMenuItem = (CheckMenuItem) menuItem;
+            if (!checkMenuItem.isSelected()) {
+                tastes.add(checkMenuItem.getText());
+            }
+        }
+
+        return tastes;
+    }
+
+    private ArrayList<Integer> getAllChosenHardnesses() {
+        ArrayList<Integer> hardnesses = new ArrayList<>();
+
+        for (MenuItem menuItem : hardnessFilterMenu.getItems()) {
+            CheckMenuItem checkMenuItem = (CheckMenuItem) menuItem;
+            if (checkMenuItem.isSelected()) {
+                hardnesses.add(Integer.valueOf(checkMenuItem.getText()));
+            }
+        }
+
+        return hardnesses;
     }
 
     @FXML
@@ -136,15 +246,18 @@ public class TabaccoListController implements Controller {
         Tabacco selectedTabacco = tabaccoListTable.getSelectionModel().getSelectedItem();
 
         if (selectedTabacco != null) {
-            TabaccoDAO.delete(selectedTabacco);
-            refresh(TabaccoDAO.getTabaccoList(Utils.getTobaccoEnumByName(selectedTabacco.getTabaccoName())));
+            boolean okClicked = fxMain.openDeleteConfirmDialog();
+            if (okClicked) {
+                TabaccoDAO.delete(selectedTabacco);
+                refreshFromButtons();
+            }
         } else {
             // Ничего не выбрано.
             Alert alert = new Alert(Alert.AlertType.WARNING);
             alert.initOwner(FxMain.getPrimaryStage());
-            alert.setTitle("No selection");
-            alert.setHeaderText("No tabacco selected");
-            alert.setContentText("Please select tabacco in the table.");
+            alert.setTitle("Нет выделеннего объекта");
+            alert.setHeaderText("Не выбран ни один табак");
+            alert.setContentText("Пожалуйста, выберете табак из таблицы.");
 
             alert.showAndWait();
         }
@@ -158,16 +271,16 @@ public class TabaccoListController implements Controller {
             boolean okClicked = fxMain.openTabaccoEditDialog(selectedTabacco);
             if (okClicked) {
                 TabaccoDAO.update(selectedTabacco);
-                refresh(TabaccoDAO.getTabaccoList(Utils.getTobaccoEnumByName(selectedTabacco.getTabaccoName())));
+                refreshFromButtons();
             }
 
         } else {
             // Ничего не выбрано.
             Alert alert = new Alert(Alert.AlertType.WARNING);
             alert.initOwner(FxMain.getPrimaryStage());
-            alert.setTitle("No selection");
-            alert.setHeaderText("No tabacco selected");
-            alert.setContentText("Please select tabacco in the table.");
+            alert.setTitle("Нет выделеннего объекта");
+            alert.setHeaderText("Не выбран ни один табак");
+            alert.setContentText("Пожалуйста, выберете табак из таблицы.");
 
             alert.showAndWait();
         }
@@ -176,23 +289,22 @@ public class TabaccoListController implements Controller {
     @FXML
     private void handleNewTabacco() {
         if (!tabaccoAddNameField.getText().equals("")) {
-            Tabacco tempTabacco = Utils.castTabaccoInSpecificClass(tabaccoAddNameField.getText());
+            Tabacco tempTabacco = Utils.convertStringInTobacco(tabaccoAddNameField.getText());
             boolean okClicked = fxMain.openTabaccoEditDialog(tempTabacco);
             if (okClicked) {
                 TabaccoDAO.add(tempTabacco);
-                refresh(TabaccoDAO.getTabaccoList(Utils.getTobaccoEnumByName(tempTabacco.getTabaccoName())));
+                refreshFromButtons();
             }
         }
 
         else {
             Alert alert = new Alert(Alert.AlertType.WARNING);
             alert.initOwner(FxMain.getPrimaryStage());
-            alert.setTitle("No tabacco name entered");
-            alert.setHeaderText("No tabacco name entered");
-            alert.setContentText("Please enter tabacco name in text field.");
+            alert.setTitle("Нет названия");
+            alert.setHeaderText("Не введено название марки табака");
+            alert.setContentText("Пожалуйста, введите название марки табака в текстовое поле.");
 
             alert.showAndWait();
         }
-
     }
 }
